Quranic-Analysis/general_quantitative_analysis/requirements.txt
pandas>=2.0.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.11.0
arabic-reshaper>=3.0.0
python-bidi>=0.6.0
nltk>=3.7.0
jupyter>=1.0.0
ipykernel>=6.0.0

